+++ packages/html/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @web-widget/html
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @web-widget/schema@0.1.1
+
 ## 0.1.0
 
 ### Patch Changes